+++ lib/phoenix_live_view/engine.ex
@@ -722,8 +722,8 @@ defmodule Phoenix.LiveView.Engine do
     to_safe(ast, line, extra_clauses)
   end
 
-  defp line_from_expr({_, meta, _}) when is_list(meta), do: Keyword.get(meta, :line)
-  defp line_from_expr(_), do: nil
+  defp line_from_expr({_, meta, _}) when is_list(meta), do: Keyword.get(meta, :line, 0)
+  defp line_from_expr(_), do: 0
 
   # We can do the work at compile time
   defp to_safe(literal, _line, _extra_clauses)