+++ test/phoenix_live_view/integrations/live_view_test.exs
@@ -3,9 +3,6 @@ defmodule Phoenix.LiveView.LiveViewTest do
   use Phoenix.ConnTest
 
   import Phoenix.LiveViewTest
-  import Phoenix.LiveView.Helpers
-
-  alias Phoenix.LiveView
   alias Phoenix.LiveViewTest.{Endpoint, DOM, ClockLive, ClockControlsLive, LayoutView}
 
   @endpoint Endpoint
@@ -524,65 +521,6 @@ defmodule Phoenix.LiveView.LiveViewTest do
                :ok = GenServer.call(view.pid, {:dynamic_child, :static})
                catch_exit(render(view))
              end) =~ "duplicate LiveView id: \"static\""
-    end
-  end
-
-  describe "redirects" do
-    test "redirect from root thru sync call", %{conn: conn} do
-      assert {:ok, view, _} = live(conn, "/thermo")
-
-      assert_redirect(view, "/path", fn ->
-        assert render_click(view, :redir, "/path") == {:error, {:redirect, %{to: "/path"}}}
-      end)
-
-      assert_remove(view, {:redirect, "/path"})
-    end
-
-    test "redirect from root thru async call", %{conn: conn} do
-      assert {:ok, view, _} = live(conn, "/thermo")
-
-      assert_redirect(view, "/async", fn ->
-        send(view.pid, {:redir, "/async"})
-      end)
-    end
-
-    test "raises from child", %{conn: conn} do
-      {:ok, thermo_view, _html} = live(conn, "/thermo")
-      GenServer.call(thermo_view.pid, {:set, :nest, []})
-      assert clock_view = find_child(thermo_view, "clock")
-
-      send(
-        clock_view.pid,
-        {:run,
-         fn socket ->
-           {:noreply, LiveView.live_redirect(socket, to: "/anywhere")}
-         end}
-      )
-
-      assert_remove(clock_view, {%ArgumentError{message: msg}, _stack})
-      assert msg =~ "cannot invoke live_redirect/2 from a child LiveView"
-    end
-  end
-
-  describe "live_link" do
-    test "forwards dom attribute options" do
-      dom =
-        live_link("next", to: "/", class: "btn btn-large", data: [page_number: 2])
-        |> Phoenix.HTML.safe_to_string()
-
-      assert dom =~ ~s|class="btn btn-large"|
-      assert dom =~ ~s|data-page-number="2"|
-    end
-
-    test "overwrites reserved options" do
-      dom =
-        live_link("next", to: "page-1", href: "page-2", data: [phx_live_link: "other"])
-        |> Phoenix.HTML.safe_to_string()
-
-      assert dom =~ ~s|href="page-1"|
-      refute dom =~ ~s|href="page-2"|
-      assert dom =~ ~s|data-phx-live-link="push"|
-      refute dom =~ ~s|data-phx-live-link="other"|
     end
   end